// github/types/paths/gists/{gist_id}/comments/{comment_id}.types.ts
export type gistsGetComment = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: gistsGetComment_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: gist_comment;
          };
        };
        examples: {
          default: unknown;
        };
      };
      304: not_modified;
      403: forbidden_gist;
      404: not_found;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Updates a comment on a gist.
 *
 * This endpoint supports the following custom media types. For more information, see "[Media types](https://docs.github.com/rest/using-the-rest-api/getting-started-with-the-rest-api#media-types)."
 *
 * - **`application/vnd.github.raw+json`**: Returns the raw markdown. This is the default if you do not pass any specific media type.
 */
export type gistsUpdateComment = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: gistsUpdateComment_Path;
    headers: never;
    cookie: never;
    body: {
      /**
       * The comment text.
       * @example "Body of the attachment"
       */
      body: string;
    };
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: gist_comment;
          };
        };
        examples: {
          default: unknown;
        };
      };
      404: not_found;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type gistsDeleteComment = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: gistsDeleteComment_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      304: not_modified;
      403: forbidden;
      404: not_found;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type gistsGetComment_Path = { gist_id: string; comment_id: number };

export type gistsUpdateComment_Path = { gist_id: string; comment_id: number };

export type gistsDeleteComment_Path = { gist_id: string; comment_id: number };